# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: RC Setup Hub frontend API
  version: 1.0.0
  description: 'Phase 11 contract for the separately developed frontend. JSON API
    uses backend-managed HttpOnly rc_session cookies. Google OAuth -> backend callback
    -> session cookie -> 303 APP_URL. Frontend requests require credentials: "include"
    and an exact ALLOWED_ORIGINS entry. No Google tokens are delivered to the frontend;
    never use localStorage auth. All /api/v1 endpoints require a session. Error bodies
    are {"error":"human-readable message"}. Browser mutations require a trusted configured
    Origin; admin HTML routes are separate and not documented here.'
servers:
  - url: http://localhost:8080
    description: Local backend; override this URL for deployed environments.
security:
  - CookieAuth: []
...
paths:
...
  /auth/google:
    get:
      operationId: startGoogleLogin
      summary: Start Google login
      tags:
        - Authentication
      responses:
        "302":
          description: Success
          headers:
            Location:
              description: Redirect destination.
              schema:
                type: string
        "500":
          $ref: '#/components/responses/InternalError'
      description: 'Navigate the browser here; do not use fetch for the Google redirect
        flow. Identity scopes: openid email profile. Backend creates a short-lived
        single-use browser-bound state, nonce and PKCE verifier, then redirects to
        Google. Arbitrary return_to/redirect parameters are ignored.'
      security: []
  /auth/google/callback:
    get:
      operationId: completeGoogleLogin
      summary: Complete Google login
      tags:
        - Authentication
      responses:
        "303":
          description: Success
          headers:
            Location:
              description: Redirect destination.
              schema:
                type: string
            Set-Cookie:
              description: Creates the backend-owned HttpOnly rc_session cookie. Tokens
                and session data are never appended to Location.
              schema:
                type: string
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "409":
          $ref: '#/components/responses/Conflict'
        "500":
          $ref: '#/components/responses/InternalError'
      description: Google redirects here. Backend verifies OAuth state, nonce and
        Google identity, upserts the user by stable Google subject and creates its
        own session. Redirects with 303 to trusted server-configured APP_URL. Never
        redirects to a request-supplied destination. Frontend receives no Google tokens.
        The callback requires the browser-bound OAuth state cookie from login start.
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
          description: Google authorization code; required on successful authorization.
        - name: state
          in: query
          required: true
          schema:
            type: string
          description: Must match the unexpired single-use state and browser cookie.
        - name: error
          in: query
          required: false
          schema:
            type: string
          description: Google authorization failure; returns 400.
      security: []
  /api/v1/auth/logout:
    post:
      operationId: logout
      summary: Invalidate the current session
      tags:
        - Authentication
      responses:
        "204":
          description: Success
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "500":
          $ref: '#/components/responses/InternalError'
      description: Revokes the server-side session and clears the HttpOnly cookie.
        No JSON response body.
  /api/v1/me:
    get:
      operationId: getMe
      summary: Get current profile
      tags:
        - Users
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalError'
    patch:
      operationId: patchMe
      summary: Change current nickname
      tags:
        - Users
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "403":
          $ref: '#/components/responses/Forbidden'
        "409":
          $ref: '#/components/responses/Conflict'
        "415":
          $ref: '#/components/responses/UnsupportedMediaType'
        "500":
          $ref: '#/components/responses/InternalError'
      description: Trims whitespace; rejects empty/overlong nickname. Case-insensitive
        duplicate nickname returns 409.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchMe'
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: rc_session
      description: Backend-issued HttpOnly cookie, never a frontend-managed token.
        Local HTTP uses SameSite=Lax, Secure=false. Production uses Secure=true; SameSite=None
        requires Secure for genuinely cross-site deployments. Sessions are in-memory,
        expire after 24h and are revoked on logout; server restart invalidates them.
  schemas:
    APIError:
      type: object
      additionalProperties: false
      properties:
        error:
          type: string
      required:
        - error
      description: Human-readable error; internal database/authentication details
        are not returned.
...
    User:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        nickname:
          type: string
          minLength: 1
          maxLength: 64
          description: Trimmed, non-empty nickname, case-insensitively unique. No
            null bytes.
        avatar_url:
          type:
            - string
            - "null"
        created_at:
          type: string
          format: date-time
      required:
        - id
        - email
        - nickname
        - avatar_url
        - created_at
      description: Current authenticated application profile. Email is exposed only
        to this user.
...
    PatchMe:
      type: object
      additionalProperties: false
      properties:
        nickname:
          type: string
          minLength: 1
          maxLength: 64
          description: Trimmed, non-empty nickname, case-insensitively unique. No
            null bytes.
      required:
        - nickname
      description: 'Nickname only. Unknown fields are rejected. JSON body limit: 4096
        bytes.'
...
  responses:
    BadRequest:
      description: Invalid input, query, UUID, JSON or OAuth state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
    Unauthorized:
      description: No valid application session or Google identity verification failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
    Forbidden:
      description: Untrusted browser mutation origin.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
...
    Conflict:
      description: Uniqueness, relationship state or concurrent update conflict.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
    UnsupportedMediaType:
      description: Content-Type must be application/json.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
    InternalError:
      description: Unexpected failure; no internal error details.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'